requirements.txt
elasticsearch
fuzzywuzzy
python-Levenshtein
streamlit
python-dotenv